+++ packages/network/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support factory on `cacheCustomKey` method.
 - Fix `cache` method to return `T` instead of `dynamic`.
 - Fix `_factory` method should take response after `_dataTransformer` done.
+- Improve cache handling in `GetBondApiRequest`.
 
 ## 0.0.10
 ### Added